requirements.txt
asgiref==3.8.1
Django~=5.1.2
psycopg2-binary==2.9.10
python-decouple==3.8
sqlparse==0.5.1
typing_extensions==4.12.2
django-scheduler~=0.10.1
pytz~=2024.2